Added random cropping

+++ Implementation/dataset_DSTL.py
@@ -1,26 +1,24 @@
-from __future__ import print_function, division
-import os
-import torch
+import os, random, cv2, torch
+from pathlib import Path
 import pandas as pd
-import cv2
 from torch.utils.data import Dataset
 import numpy as np
 import tifffile as tiff
-from data_import import generate_mask_for_image_and_class, _get_polygon_list, _get_image_names
+from data_import import generate_mask_for_image_and_class, _get_polygon_list
 
 
-class myDatasetClass(Dataset):
+class datasetDSTL(Dataset):
     """Face Landmarks dataset."""
 
-    def __init__(self, dir_path, inputPath, channel='rgb-g' ,res=(0,0) ,transform=None):
+    def __init__(self, dir_path, inputPath, channel='rgb' ,res=(0,0), crange=(0.1, 0.4)):
         """
         Args:
             dir_path (string):  Working Directory.
             inputPath (string): Input directory with all the images.
             channel (string):   Channels used for the input, eiter 'gray', 'rgb', 'rgb-g' (for rgb and gray) or 'all' (tbd).
-            res (tuple/list):    Resolution for the input images and masks. If 'all' is selected must be <= (132, 133) pixel.
-                                        If resolution == (0, 0) then it will remains unchanges for the data
-            transform (callable, optional): Optional transform to be applied on a sample.
+            res (tuple/list):   Resolution for the input images and masks. If 'all' is selected must be <= (132, 133) pixel.
+                                If resolution == (0, 0) then it will remains unchanges for the data
+            crange (tuple):     Defines the range of how strong the random crop for images can be. The direction is also chosen randomly
         """        
         
         inpDir = str(dir_path)+str(inputPath)
@@ -31,7 +29,7 @@ class myDatasetClass(Dataset):
         self.dir_path = dir_path
         self.imgIDs = imgIDs
         self.channel = channel
-        self.transform = transform
+        self.crange = crange
         df = pd.read_csv(inpDir + 'train_wkt_v4.csv')
         gs = pd.read_csv(inpDir + 'grid_sizes.csv', names=['ImageId', 'Xmax', 'Ymin'], skiprows=1)
         newInputs = []
@@ -56,7 +54,7 @@ class myDatasetClass(Dataset):
                                 grayImg = cv2.resize(grayImg, res)
                             rgbgImage = cv2.merge([rgbImage,grayImg])
                             newInputs.append(self.saveNewImage('\\rgb-g\\',rgbgImage, imageId))
-            print('Processing Images: '+str((idx/len(imgIDs))*100)+'%')
+                print('Processing Images: '+str((idx/len(imgIDs))*100)+'%')
         else:
             if(channel=='rgb'):
                 newInputs = [x for x in inputs if (not x.endswith('_P.tif') and not x.endswith('_M.tif') and not x.endswith('_A.tif'))]
@@ -79,7 +77,9 @@ class myDatasetClass(Dataset):
                 #polygons.append(d)
                 mask = generate_mask_for_image_and_class(res,imageId,classType,gs,df)
                 filename = str(dir_path)+'\\masks\\'+str(imageId)+'-'+str(classType)+'-'+str(self.res[0])+'x'+str(self.res[1])+'.png'
-                cv2.imwrite(filename,mask*255)
+                my_file = Path(filename)
+                if not my_file.is_file():
+                    cv2.imwrite(filename,mask*255)
                 masksNames.append(filename)
             masks.append( masksNames )
             print('Processing Masks: '+str((idx/len(imgIDs))*100)+'%')
@@ -90,7 +90,9 @@ class myDatasetClass(Dataset):
 
     def saveNewImage(self, path, img, imageId):
             filename = str(self.dir_path)+str(path)+str(imageId)+'-'+str(self.res[0])+'x'+str(self.res[1])+'.png'
-            cv2.imwrite(filename , img)
+            my_file = Path(filename)
+            if not my_file.is_file():
+                cv2.imwrite(filename , img)
             return filename
 
     def getIDsAndFiles(self, inpDir):
@@ -148,16 +150,43 @@ class myDatasetClass(Dataset):
     def __len__(self):
         return len(self.inputs)
 
+    def randomCrop(self, image, dir, strength):
+
+        h, w = image.shape[:2]
+        x = np.cos((dir*np.pi)/180)
+        y = np.sin((dir*np.pi)/180)
+        x_new = (x/(np.abs(x)+np.abs(y)))*strength
+        y_new = (y/(np.abs(x)+np.abs(y)))*strength
+
+        ox = int(x_new*w)
+        oy = int(y_new*h)
+        non = lambda s: s if s<0 else None
+        mom = lambda s: max(0,s)
+        shift_img = np.zeros_like(image)
+        shift_img[mom(oy):non(oy), mom(ox):non(ox)] = image[mom(-oy):non(-oy), mom(-ox):non(-ox)]
+        #cv2.imshow(" ",shift_img)
+        #cv2.waitKey(0)
+        return shift_img
+
+    def toTensor(self, image):
+        # is this necessary? The image can contain also other channels ...
+
+        # swap color axis because
+        # numpy image: H x W x C
+        # torch image: C X H X W
+        # image = image.transpose((2, 0, 1))
+        return image
+
     def __getitem__(self, idx):
+        strength = random.uniform(self.crange[0],self.crange[1])
+        dir = np.random.randint(0,360) 
         imageId = self.imgIDs[idx]
-        image = cv2.imread(self.inputs[idx])
+        image = self.toTensor(self.randomCrop(cv2.imread(self.inputs[idx],cv2.IMREAD_UNCHANGED),dir,strength))
         masks = self.masks[idx]
         masksImgs = []
-        for maskFile in masks:
-            masksImgs.append(cv2.imread(maskFile))
-        item = {'image': image, 'masks': masksImgs}
 
-        if self.transform:
-            item = self.transform(item)
+        for maskFile in masks:
+            masksImgs.append(self.toTensor(self.randomCrop(cv2.imread(maskFile),dir,strength)))
+        item = {'image': self.toTensor(image), 'masks': masksImgs}
 
         return item
